requirements.txt
requests
beautifulsoup4
psycopg2-binary
python-telegram-bot>=22.1
python-dotenv
